Further improvements, bug fixes, legend and hover
