+++ frontend/src/modules/PvtRefactored/utils/PvtPlotBuilder.ts
@@ -1,3 +1,4 @@
+import { EnsembleIdent } from "@framework/EnsembleIdent";
 import { ColorSet } from "@lib/utils/ColorSet";
 import { Size2D } from "@lib/utils/geometry";
 import { Figure, makeSubplots } from "@modules/_shared/Figure";
@@ -11,9 +12,11 @@ import {
     PRESSURE_DEPENDENT_VARIABLE_TO_DISPLAY_NAME,
     PhaseType,
     PressureDependentVariable,
+    PvtTableCollection,
 } from "../typesAndEnums";
 
 type TracePointData = {
+    ratio: number;
     pressure: number;
     dependentVariableValue: number;
 };
@@ -22,13 +25,21 @@ export class PvtPlotBuilder {
     private _pvtDataAccessor: PvtDataAccessor;
     private _figure: Figure | null = null;
     private _numPlots = 0;
+    private _makeEnsembleDisplayNameFunc: (ensemble: EnsembleIdent) => string;
 
-    constructor(pvtDataAccessor: PvtDataAccessor) {
+    constructor(pvtDataAccessor: PvtDataAccessor, makeEnsembleDisplayNameFunc: (ensemble: EnsembleIdent) => string) {
         this._pvtDataAccessor = pvtDataAccessor;
+        this._makeEnsembleDisplayNameFunc = makeEnsembleDisplayNameFunc;
     }
 
-    makeLayout(selectedPlots: PressureDependentVariable[], size: Size2D): void {
-        this._numPlots = selectedPlots.length;
+    makeLayout(phase: PhaseType, dependentVariables: PressureDependentVariable[], size: Size2D): void {
+        const adjustedDependentVariables = dependentVariables.filter((el) => {
+            if (phase === PhaseType.WATER) {
+                return el !== PressureDependentVariable.FLUID_RATIO;
+            }
+            return true;
+        });
+        this._numPlots = adjustedDependentVariables.length;
         const numRows = Math.ceil(this._numPlots / 2);
         const numCols = Math.min(this._numPlots, 2);
         const subplotTitles: string[] = [];
@@ -40,7 +51,7 @@ export class PvtPlotBuilder {
                 const titleIndex = (numRows - row) * numCols + (col - 1);
                 const axisIndex = (row - 1) * numCols + col;
 
-                const dependentVariable = selectedPlots.at(titleIndex) ?? null;
+                const dependentVariable = adjustedDependentVariables.at(titleIndex) ?? null;
                 if (!dependentVariable) {
                     subplotTitles.push("");
                     continue;
@@ -67,13 +78,74 @@ export class PvtPlotBuilder {
             sharedXAxes: false,
             sharedYAxes: false,
             showGrid: true,
-            margin: { t: 30, b: 40, l: 40, r: 10 },
+            margin: { t: 30, b: 40, l: 60, r: 10 },
             subplotTitles: subplotTitles,
         });
 
         for (const patch of patches) {
             this._figure.updateLayout(patch);
         }
+    }
+
+    private makeHoverTemplate(
+        dependentVariable: PressureDependentVariable,
+        ratios: number[],
+        pvtNum: number,
+        phase: PhaseType,
+        ensembleIdent: EnsembleIdent,
+        realization: number
+    ): string[] {
+        const nameY = PRESSURE_DEPENDENT_VARIABLE_TO_DISPLAY_NAME[dependentVariable];
+        const ensembleDisplayName = this._makeEnsembleDisplayNameFunc(ensembleIdent);
+
+        return ratios.map((ratio) => {
+            if (phase === PhaseType.OIL) {
+                return `Pressure: <b>%{x}</b><br>${nameY}: <b>%{y}</b><br>Rs: <b>${ratio}</b><br>PVTNum: <b>${pvtNum}</b><br>Ensemble: <b>${ensembleDisplayName}</b> Realization: <b>${realization}</b>`;
+            } else if (phase === PhaseType.GAS) {
+                return `Pressure: <b>%{x}</b><br>${nameY}: <b>%{y}</b><br>Rv: <b>${ratio}</b><br>PVTNum: <b>${pvtNum}</b><br>Ensemble: <b>${ensembleDisplayName}</b> Realization: <b>${realization}</b>`;
+            }
+            return `Pressure: <b>%{x}</b><br>${nameY}: <b>%{y}</b><br>PVTNum: <b>${pvtNum}</b><br>Ensemble: <b>${ensembleDisplayName}</b> Realization: <b>${realization}</b>`;
+        });
+    }
+
+    private makeLegendTitle(colorBy: ColorBy) {
+        if (!this._figure) {
+            throw new Error("Layout not set");
+        }
+
+        let legendTitle = "Ens - Real";
+        if (colorBy === ColorBy.PVT_NUM) {
+            legendTitle = "PVTNum";
+        }
+
+        this._figure.updateLayout({
+            legend: {
+                title: {
+                    text: legendTitle,
+                },
+                orientation: "v",
+            },
+        });
+    }
+
+    private makeColorsArray(
+        colorBy: ColorBy,
+        colorSet: ColorSet,
+        pvtNumsLength: number,
+        tableCollectionsLength: number
+    ): string[] {
+        const colors: string[] = [];
+        colors.push(colorSet.getFirstColor());
+        if (colorBy === ColorBy.PVT_NUM) {
+            for (let i = 1; i < pvtNumsLength; i++) {
+                colors.push(colorSet.getNextColor());
+            }
+        } else {
+            for (let i = 1; i < tableCollectionsLength; i++) {
+                colors.push(colorSet.getNextColor());
+            }
+        }
+        return colors;
     }
 
     makeTraces(
@@ -89,29 +161,8 @@ export class PvtPlotBuilder {
 
         const tableCollections = this._pvtDataAccessor.getTableCollections();
 
-        let legendTitle = "";
-        const colors: string[] = [];
-        colors.push(colorSet.getFirstColor());
-        if (colorBy === ColorBy.PVT_NUM) {
-            for (let i = 1; i < pvtNums.length; i++) {
-                colors.push(colorSet.getNextColor());
-            }
-            legendTitle = "PVTNum";
-        } else {
-            for (let i = 1; i < tableCollections.length; i++) {
-                colors.push(colorSet.getNextColor());
-            }
-            legendTitle = "Ens + Real";
-        }
-
-        this._figure.updateLayout({
-            legend: {
-                title: {
-                    text: legendTitle,
-                },
-                orientation: "v",
-            },
-        });
+        this.makeLegendTitle(colorBy);
+        const colors = this.makeColorsArray(colorBy, colorSet, pvtNums.length, tableCollections.length);
 
         let collectionIndex = 0;
         let pvtNumIndex = 0;
@@ -170,6 +221,7 @@ export class PvtPlotBuilder {
                                 }
                             }
                             tracePointDataArray.push({
+                                ratio: table.ratio[i],
                                 pressure: table.pressure[i],
                                 dependentVariableValue,
                             });
@@ -186,16 +238,24 @@ export class PvtPlotBuilder {
                         const row = Math.ceil(this._numPlots / 2) - Math.floor(i / 2);
                         const col = (i % 2) + 1;
                         const borderTracePoints: TracePointData[] = [];
-                        for (const [groupValue, tracePointDataArray] of dependentVariableMap) {
+                        for (const [, tracePointDataArray] of dependentVariableMap) {
                             const trace: Partial<PlotData> = {
                                 x: tracePointDataArray.map((el) => el.pressure),
                                 y: tracePointDataArray.map((el) => el.dependentVariableValue),
                                 mode: "lines+markers",
-                                name: `${dependentVariable} ${groupValue}`,
                                 line: {
                                     color,
                                 },
+                                name: "",
                                 showlegend: false,
+                                hovertemplate: this.makeHoverTemplate(
+                                    dependentVariable,
+                                    tracePointDataArray.map((el) => el.ratio),
+                                    table.pvtnum,
+                                    phase,
+                                    tableCollection.ensembleIdent,
+                                    tableCollection.realization
+                                ),
                             };
 
                             this._figure.addTrace(trace, row, col);
@@ -224,10 +284,9 @@ export class PvtPlotBuilder {
                             if (colorBy === ColorBy.PVT_NUM) {
                                 traceLegendName = table.pvtnum.toString();
                             } else {
-                                traceLegendName =
-                                    tableCollection.ensembleIdent.getEnsembleName() +
-                                    " - " +
-                                    tableCollection.realization;
+                                traceLegendName = `${this._makeEnsembleDisplayNameFunc(
+                                    tableCollection.ensembleIdent
+                                )} - ${tableCollection.realization}`;
                             }
 
                             this._figure.addTrace({
@@ -240,7 +299,6 @@ export class PvtPlotBuilder {
                                 },
                                 showlegend: true,
                                 visible: "legendonly",
-                                legendgroup: legendTitle,
                             });
                         }
